refactor: fix dependency and structural violations
- Use RateLimitStore Protocol instead of concrete type in app.py
- Move WorkQueue Protocol to analysis/subagents/types.py
- Rename _resolve_evidence_paths to public resolve_evidence_paths
- Thread evaluators_dir through PromptContext instead of calling
  default_paths() in builder.py
- Import from services.ports instead of adapters in filesystem.py
