+++ src/quodeq/analysis/runner.py
@@ -91,6 +91,7 @@ def _build_dimension_prompt(
             source_file_count=config.source_file_count,
             dimensions_data=ctx.dimensions_data,
             standards_dir=config.standards_dir,
+            evaluators_dir=config.evaluators_dir,
             manifest=config.manifest,
             target=config.target,
             work_dir=config.work_dir or config.src,
