+++ src/quodeq/analysis/subagents/_consolidated.py
@@ -83,6 +83,7 @@ def _build_prompt(config: "RunConfig", dimensions: list[str], ctx: Any) -> str:
             source_file_count=config.source_file_count,
             dimensions_data=ctx.dimensions_data,
             standards_dir=config.standards_dir,
+            evaluators_dir=config.evaluators_dir,
             manifest=config.manifest,
             target=config.target,
             work_dir=config.work_dir or config.src,
